old/callico/tests-requirements.txt
openpyxl==3.1.5
pytest-django==4.8.0
pytest-lazy-fixtures==1.1.1
pytest-mock==3.14.0
pytest-responses==0.5.1
selenium==4.24.0
